Update MarketView and convert prices to BigDecimal Fill in the tick method for MarketView - it is slightly fragile at the moment, and reliant on Orders being uniquely represented. This is both to work out whether the Order was a user's buy or sell, and also to identify the Order and remove it from the MarketView's list of outstanding orders. I also went through and used BigDecimal in place of integers for every mention of price. I think I converted it all right, but Java's lack of operator overloading makes it tricky to be sure! Possible source of bugs in particular in comparisons.

+++ src/orderBookReconstructor/BuyOrder.java
@@ -1,12 +1,12 @@
 package orderBookReconstructor;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
-import java.util.Comparator;
 
 import testHarness.StockHandle;
 
 public class BuyOrder extends Order {
-	public BuyOrder(StockHandle stock, Timestamp timePlaced, int price, int volume) {
+	public BuyOrder(StockHandle stock, Timestamp timePlaced, BigDecimal price, int volume) {
 		super(stock, timePlaced, price, volume);
 	}
