+++ src/orderBookReconstructor/Match.java
@@ -1,11 +1,13 @@
 package orderBookReconstructor;
 
+import java.math.BigDecimal;
+
 public class Match {
 	public final BuyOrder buyOrder;
 	public final SellOrder sellOrder;
 	public final int quantity;
-	public final int price;
-	public Match(BuyOrder buyOrder, SellOrder sellOrder, int quantity, int price) {
+	public final BigDecimal price;
+	public Match(BuyOrder buyOrder, SellOrder sellOrder, int quantity, BigDecimal price) {
 		this.buyOrder = buyOrder;
 		this.sellOrder = sellOrder;
 		this.quantity = quantity;
